pyproject.toml
[project]
name = "graphrag-lab"
version = "0.1.0"
description = "A structured experiment framework for comparing GraphRAG implementations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "tiktoken>=0.7",
    "pandas>=2.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
microsoft = ["graphrag>=3.0"]
lightrag = ["lightrag-hku>=1.0"]
nano = ["nano-graphrag>=0.1"]
fast = ["fast-graphrag>=0.1"]
neo4j = ["neo4j-graphrag[openai,experimental]>=1.14", "neo4j>=5.0"]
datastax = ["langchain-graph-retriever>=0.1", "langchain-chroma>=0.2", "langchain-openai>=0.3"]
cognee = ["cognee>=0.1"]
graphiti = ["graphiti-core>=0.1"]
raganything = ["raganything>=0.1"]
hipporag = ["hipporag>=0.1"]
linearrag = ["linearrag>=0.1"]
pathrag = ["pathrag>=0.1"]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
ollama = ["ollama>=0.4"]
notebooks = ["jupyter>=1.0", "matplotlib>=3.8", "seaborn>=0.13", "plotly>=5.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.8"]
all = ["graphrag-lab[microsoft,lightrag,nano,fast,neo4j,datastax,cognee,graphiti,raganything,hipporag,linearrag,pathrag,anthropic,openai,ollama,notebooks,dev]"]

[project.scripts]
graphrag-lab = "graphrag_lab.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/graphrag_lab"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
